+++ public_website/__tests__/pages/home.test.tsx
@@ -2,9 +2,9 @@ import React from 'react'
 import { describe, expect, it } from 'vitest'
 import { axe } from 'vitest-axe'
 
-import Home, { getStaticProps } from '../src/pages'
-import { act, fireEvent, render, screen, waitFor } from '.'
-import { activePlaylistTagsFixtures, playlistOffersFixtures } from './fixtures'
+import Home, { getStaticProps } from '../../src/pages'
+import { act, fireEvent, render, screen, waitFor } from '..'
+import { activePlaylistTagsFixtures, playlistOffersFixtures } from '../fixtures'
 
 describe('Home page', () => {
   it('should pass axe accessibility tests', async () => {
